# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_orders_upload_confirmation_attachment": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file"],
        "title": "Body_orders_upload_confirmation_attachment",
        "type": "object"
      },
...
      "ConfirmationAttachmentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConfirmationAttachment"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": ["items"],
        "title": "ConfirmationAttachmentPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderCancel": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": ["version"],
        "title": "OrderCancel",
        "type": "object"
      },
...
      "OrderComplete": {
        "additionalProperties": false,
        "properties": {
          "confirmation_signer_name": {
            "maxLength": 120,
            "title": "Confirmation Signer Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": ["version", "confirmation_signer_name"],
        "title": "OrderComplete",
        "type": "object"
      },
...
      "OrderPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "OrderPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SMYou Pro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "orders_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "DRAFT",
                    "PENDING_DISPATCH",
                    "IN_PROGRESS",
                    "AWAITING_CONFIRMATION",
                    "COMPLETED",
                    "REVISION",
                    "CANCELLED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at_desc",
              "enum": ["created_at_desc", "dispatch"],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Danh sách đơn (tìm theo mã/tên khách/SĐT, lọc theo trạng thái)",
        "tags": ["orders"]
      },
      "post": {
        "operationId": "orders_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tạo đơn nháp",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "operationId": "orders_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chi tiết đơn",
        "tags": ["orders"]
      },
      "patch": {
        "operationId": "orders_update",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | ORDER_NOT_DRAFT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sửa thông tin đơn nháp",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/cancel": {
      "post": {
        "operationId": "orders_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCancel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | INVALID_TRANSITION | GUARD_FAILED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Huỷ đơn",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/complete": {
      "post": {
        "operationId": "orders_complete",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderComplete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | INVALID_TRANSITION | GUARD_FAILED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hoàn tất đơn",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/confirmation-attachments": {
      "get": {
        "operationId": "orders_list_confirmation_attachments",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationAttachmentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Danh sách ảnh phiếu xác nhận",
        "tags": ["orders"]
      },
      "post": {
        "operationId": "orders_upload_confirmation_attachment",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_orders_upload_confirmation_attachment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationAttachment"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "422": {
            "description": "problem+json — INVALID_FILE_TYPE | FILE_TOO_LARGE | UNSUPPORTED_MEDIA_TYPE"
          }
        },
        "summary": "Tải ảnh phiếu xác nhận",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/contact": {
      "patch": {
        "operationId": "orders_update_contact",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderContactUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | ORDER_NOT_SUBMITTED | ORDER_LOCKED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sửa liên hệ/mô tả sau khi gửi",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/history": {
      "get": {
        "operationId": "orders_history",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lịch sử thay đổi trạng thái của đơn",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/lines": {
      "post": {
        "operationId": "orders_add_line",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderLineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | ORDER_NOT_DRAFT"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thêm dòng hàng vào đơn nháp",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/lines-after-submit": {
      "post": {
        "operationId": "orders_add_line_after_submit",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderLineCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | ORDER_NOT_SUBMITTED | ORDER_LOCKED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thêm dòng hàng sau khi gửi",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{order_id}/lines-after-submit/{line_id}": {
      "patch": {
        "operationId": "orders_update_line_after_submit",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "line_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Line Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderLineUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | ORDER_NOT_SUBMITTED | ORDER_LOCKED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sửa dòng hàng sau khi gửi",
        "tags": ["orders"]
      }
    },
...
  }
}